requirements.txt
flask
gunicorn
wheel
psycopg2-binary
decouple